api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
motor
python-dotenv
langchain
langchain-community
langgraph
langchain-ollama